+++ MinecraftClient/MinecraftClient.csproj
@@ -77,7 +77,6 @@
     <Compile Include="ChatBots\Mailer.cs" />
     <Compile Include="Mapping\CubeFromWorld.cs" />
     <Compile Include="Mapping\Material2Tool.cs" />
-    <Compile Include="ChatBots\MineCube.cs" />
     <Compile Include="ChatBots\ReplayCapture.cs" />
     <Compile Include="Commands\DropItem.cs" />
     <Compile Include="Commands\Entitycmd.cs" />
